requirements.txt
Django==1.11.6
jsmin==2.2.2
libsass==0.13.3
pytz==2017.3
six==1.11.0
webassets==0.12.1
